requirements.txt
anthropic>=0.40.0
requests>=2.31.0
Pillow>=10.0.0
opencv-python-headless>=4.8.0
numpy>=1.24.0
pyyaml>=6.0
rembg[gpu]>=2.0.50
onnxruntime-gpu>=1.16.0
python-dotenv>=1.0.0
tqdm>=4.66.0
imageio>=2.31.0
